Rename presenters folder
If we were following the pattern in services they really should be in `app/presenters/billing/supplementary`! But we need to create the same data structure for creating bill runs and their events for all types of bill run so actually we can rename the folder as `/billing`.
